chore: move gateway from internal/ to pkg/ (#1368)
