+++ pkg/gateway/doc.go
@@ -1,0 +1,2 @@
+// Package gateway provides implementations to send data through the transport (e.g. headers).
+package gateway
